change React's folder structure
