+++ frontend/src/Pages/CreateFormJose/CreateForm.module.css
@@ -13,7 +13,6 @@ button {
 }
 
 .fixedHeader {
-    position: ;
     top: 0 !important;
     left: 0 !important;
     width: 100%;
