+++ frontend/src/Pages/CreateFormJose/CreateFormJose.jsx
@@ -1,6 +1,6 @@
 import { useFieldArray, useForm } from 'react-hook-form';
 import { useEffect, useState } from 'react';
-import { api } from '../../Utils/api.js';
+import { api } from '../../utils/api.js';
 import { useQuery, useQueryClient } from 'react-query';
 import { useNavigate, useParams, Link } from 'react-router-dom';
 import styles from './CreateForm.module.css'
@@ -131,11 +131,17 @@ export const EditForm = () => {
   const queryClient = useQueryClient();
 
   const onSubmit = (data) => {
-    api().patch(`/form/${id}`, data).then((response) => {
+    const processedData = {
+      ...data,
+      questions: data.questions.map((question) => {
+        const { _id, ...rest } = question;
+        return rest
+      })
+    };
+    api().patch(`/form/${id}`, processedData).then((response) => {
       console.log(response.data);
       queryClient.invalidateQueries('forms');
       navigate('/workspace');
-
     });
   }
   return (
