+++ frontend/src/Pages/Home/Home.module.css
@@ -10,8 +10,8 @@ body {
 }
 
 .background {
-    background: rgb(160,158,194);
-    background: linear-gradient(90deg, rgba(160,158,194,0.5410758053221288) 0%, rgba(197,197,255,0.5550814075630253) 43%, rgba(0,212,255,0.4710477941176471) 100%);
+    background: rgb(160, 158, 194);
+    background: linear-gradient(90deg, rgba(160, 158, 194, 0.5410758053221288) 0%, rgba(197, 197, 255, 0.5550814075630253) 43%, rgba(0, 212, 255, 0.4710477941176471) 100%);
     width: 100vw;
 }
 
@@ -28,7 +28,8 @@ body {
     left: 50%;
     transform: translateX(-50%);
     z-index: 999;
-    margin-top: 30px; /* add this line to create space between logo and the top */
+    margin-top: 30px;
+    /* add this line to create space between logo and the top */
 }
 
 .welcome {
@@ -41,9 +42,9 @@ body {
     height: 50vw;
     width: 95vw;
     border-radius: 20px;
-    -webkit-box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
-    -moz-box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
-    box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
+    -webkit-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+    -moz-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+    box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
 }
 
 h1 {
@@ -66,14 +67,14 @@ button {
     border-radius: 50px;
     border: 1px solid rgba(162, 156, 156, 0.433);
     transition: 0.3s;
-    -webkit-box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
-    -moz-box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
-    box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
+    -webkit-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+    -moz-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+    box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
 }
 
 button:hover {
     box-shadow: 0 0 rgba(0, 0, 0, .3);
-  }
+}
 
 .flow {
     font-family: "Rubik", sans-serif;
@@ -92,45 +93,45 @@ button:hover {
 }
 
 @keyframes myAnim {
-	0% {
-		animation-timing-function: ease-in;
-		opacity: 0;
-		transform: translateX(-250px);
-	}
+    0% {
+        animation-timing-function: ease-in;
+        opacity: 0;
+        transform: translateX(-250px);
+    }
 
-	38% {
-		animation-timing-function: ease-out;
-		opacity: 1;
-		transform: translateX(0);
-	}
+    38% {
+        animation-timing-function: ease-out;
+        opacity: 1;
+        transform: translateX(0);
+    }
 
-	55% {
-		animation-timing-function: ease-in;
-		transform: translateX(-68px);
-	}
+    55% {
+        animation-timing-function: ease-in;
+        transform: translateX(-68px);
+    }
 
-	72% {
-		animation-timing-function: ease-out;
-		transform: translateX(0);
-	}
+    72% {
+        animation-timing-function: ease-out;
+        transform: translateX(0);
+    }
 
-	81% {
-		animation-timing-function: ease-in;
-		transform: translateX(-28px);
-	}
+    81% {
+        animation-timing-function: ease-in;
+        transform: translateX(-28px);
+    }
 
-	90% {
-		animation-timing-function: ease-out;
-		transform: translateX(0);
-	}
+    90% {
+        animation-timing-function: ease-out;
+        transform: translateX(0);
+    }
 
-	95% {
-		animation-timing-function: ease-in;
-		transform: translateX(-8px);
-	}
+    95% {
+        animation-timing-function: ease-in;
+        transform: translateX(-8px);
+    }
 
-	100% {
-		animation-timing-function: ease-out;
-		transform: translateX(0);
-	}
+    100% {
+        animation-timing-function: ease-out;
+        transform: translateX(0);
+    }
 }
