+++ frontend/src/Pages/Login/Login.jsx
@@ -1,0 +1,12 @@
+import styles from './Login.module.css';
+import LoginForm from './LoginForm';
+
+const Login = () => {
+    return (
+        <div className={styles.mastercontainer}>
+            <LoginForm />
+        </div>
+    );
+};
+
+export default Login;
