+++ frontend/src/Pages/Login/Login.module.css
@@ -52,9 +52,9 @@
     border-radius: 10px;
     border: 1px solid rgba(162, 156, 156, 0.433);
     transition: 0.3s;
-    -webkit-box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
-    -moz-box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
-    box-shadow: 0px 0px 5px 1px rgba(171,171,171,1);
+    -webkit-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+    -moz-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+    box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
 }
 
 .submitbuttonbutton:hover {
