+++ frontend/src/index.css
@@ -1,0 +1,24 @@
+h1 {
+    font-family: Abril Fatface;
+}
+
+button {
+    font-family: "Space Mono", monospace;
+    font-weight: 400;
+    font-style: normal;
+    font-size: 20px;
+    background-color: azure;
+    height: 70px;
+    min-width: max-content;
+    margin-top: 50px;
+    border-radius: 50px;
+    border: px solid rgba(162, 156, 156, 0.433);
+    transition: 0.3s;
+    -webkit-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+    -moz-box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+    box-shadow: 0px 0px 5px 1px rgba(171, 171, 171, 1);
+}
+
+button:hover {
+    box-shadow: 0 0 rgba(0, 0, 0, .3);
+}
